:building_construction: const: User 도메인 Service 레이어 패키지 이동 및 네이밍 변경
